refactor: move HTML-to-text stripping into base/message.py so it can be reused by imap/message.py
